pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "goproverlay"
version = "0.1.0"
description = "CLI to overlay fitness metrics and GPS track onto videos using FIT files"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
  { name = "Your Name" }
]
dependencies = [
  "typer[all]>=0.9.0",
  "moviepy>=1.0.3",
  "fitdecode>=0.10.0",
  "Pillow>=10.0.0",
  "numpy>=1.22.0"
]

[project.scripts]
goproverlay = "goproverlay.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["goproverlay"]
